+++ data/raw/sdss_180.0_0.0.jpg
