+++ data/raw/sdss_190.06253277483466_-11.614957355709242.jpg
