+++ data/raw/sdss_229.525575753922_42.7458537608544.jpg
